Add run_at support !!DROPS RAILS 2.3 SUPPORT!!
Added the support for scheduling jobs in the future with run_at.
In scalers/base.rb I've used the where on the Delayed Jobs table to get the pending jobs.

This forces the use of Rails 3.x! So Rails 2.3 support has been dropped in this commit. If you know the correct way to do this query in a Rails 2.3 and Rails 3.x way, please inform me.

Also this introduces a middleware component to check for pending jobs, and if there are any, it will call to scale up the workers.

Signed-off-by: David Westerink <davidakachaos@gmail.com>